requirements.txt
discord.py=~1.5.0
jishaku=~1.19.2.202